Module_6/MCP (non-adk)/agents/mcp_agent.py
# More complete implementation of mcp_agent.py
import asyncio
import logging
import uuid
from dotenv import load_dotenv
from google.genai import types
from google.adk.agents.llm_agent import LlmAgent
from google.adk.runners import Runner
from google.adk.sessions import InMemorySessionService
from google.adk.artifacts.in_memory_artifact_service import InMemoryArtifactService
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset, StdioServerParameters
import os
import sys


# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load environment variables if needed
load_dotenv()

async def get_tools_async():
    """Gets tools from the MCP Server."""
    logger.info("Connecting to MCP security-hub server...")

    try:
        # Connect to your existing MCP server
        tools, exit_stack = await MCPToolset.from_server(
            connection_params=StdioServerParameters(
                command='python',  # Command to run the server
                args=[
                    "./servers/server_mcp.py"  # Your existing MCP server
                ],
            )
        )

        logger.info(f"MCP Toolset created successfully with {len(tools)} tools")
        return tools, exit_stack
    except Exception as e:
        logger.error(f"Failed to connect to MCP server: {e}")
        raise

async def get_agent_async():
    """Creates an ADK Agent equipped with tools from the MCP Server."""
    try:
        tools, exit_stack = await get_tools_async()

        # Create the agent with MCP tools
        root_agent = LlmAgent(
            model='gemini-2.5-pro-preview-03-25',  # Match your model from query_MCP_ADK.py
            name='sql_analysis_assistant',
            instruction="""
            You are an expert SQL analyst working with a walmart_sales database.
            Use the `query_data` tool to run your SQL queries.
            Do not use `execute_sql_query`.
            Make sure all SQL is UPPER CASE and valid for the database.
            Return only the result of the query, nothing else.
            """,
            tools=tools,  # Provide the MCP tools to the ADK agent
        )

        return root_agent, exit_stack
    except Exception as e:
        logger.error(f"Failed to create agent: {e}")
        raise

async def run_mcp_agent(query):
    """Run the MCP agent with a given query and return the response."""
    session_service = InMemorySessionService()
    artifacts_service = InMemoryArtifactService()
    exit_stack = None

    try:
        # Create a unique session with a UUID
        session_id = f"session_{uuid.uuid4()}"
        session = session_service.create_session(
            state={},
            app_name='mcp_sql_analysis_app',
            user_id='user_1',  # Using your existing USER_ID
            session_id=session_id
        )

        logger.info(f"User Query: '{query}'")
        content = types.Content(role='user', parts=[types.Part(text=query)])

        # Get the agent with MCP tools
        root_agent, exit_stack = await get_agent_async()

        # Create runner
        runner = Runner(
            app_name='mcp_sql_analysis_app',
            agent=root_agent,
            artifact_service=artifacts_service,
            session_service=session_service,
        )

        logger.info("Running agent...")
        result_text = ""

        # Process the query
        events_async = runner.run_async(
            session_id=session.id,
            user_id=session.user_id,
            new_message=content
        )

        async for event in events_async:
            logger.debug(f"Event type: {type(event)}")
            if event.is_final_response() and event.content and event.content.parts:
                result_text = event.content.parts[0].text

        return result_text
    except Exception as e:
        logger.error(f"Error running MCP agent: {e}")
        return f"Error: {str(e)}"
    finally:
        # Clean up MCP connection
        if exit_stack:
            logger.info("Closing MCP server connection...")
            try:
                await exit_stack.aclose()
            except Exception as e:
                logger.error(f"Error closing MCP connection: {e}")
